Revert "Create Database Framework"
