+++ pytextnow/TNAPI.py
@@ -1,20 +1,29 @@
-from login import login
-from TN_objects.error import FailedRequest, AuthError, InvalidEvent
-from TN_objects.multi_media_message import MultiMediaMessage
-from TN_objects.message import Message
-from TN_objects.container import Container
-from TN_objects.contact import Contact
-from constants import *
-
+if __name__ == "__main__":
+    from login import login
+else:
+    from pytextnow.login import login
+    from pytextnow.error import FailedRequest, AuthError, InvalidEvent
+    from pytextnow.message_container import MessageContainer
+    from pytextnow.multi_media_message import MultiMediaMessage
+    from pytextnow.message import Message
 import mimetypes
 import requests
 from datetime import datetime, time
-import datetime as dt
+from dateutil.relativedelta import relativedelta
 
 import json
 from os.path import realpath, dirname, join
 import time
 import atexit
+
+MESSAGE_TYPE = 0
+MULTIMEDIA_MESSAGE_TYPE = 1
+
+SENT_MESSAGE_TYPE = 2
+RECEIVED_MESSAGE_TYPE = 1
+
+SIP_ENDPOINT = "prod.tncp.textnow.com"
+
 
 class Client:
     def __init__(self, username: str = None, cookie=None):
@@ -59,25 +68,23 @@ class Client:
                           'Chrome/88.0.4324.104 Safari/537.36 '
         }
 
-        self.session = requests.session()
-
         file.close()
 
-        atexit.register(self.on_exit)
+        def on_exit():
+            if len(self.events) == 0:
+                return
+
+            while 1:
+                for event, func in self.events:
+                    if event == "message":
+                        unread_msgs = self.get_unread_messages()
+                        for msg in unread_msgs:
+                            msg.mark_as_read()
+                            func(msg)
+
+        atexit.register(on_exit)
 
     # Functions
-    def on_exit(self):
-        if len(self.events) == 0:
-            return
-
-        while 1:
-            for event, func in self.events:
-                if event == "message":
-                    unread_msgs = self.get_unread_messages()
-                    for msg in unread_msgs:
-                        msg.mark_as_read()
-                        func(msg)
-
     def auth_reset(self, cookie=None):
         with open(self._user_sid_file, "r") as user_SIDS_file:
             user_SIDS = json.loads(user_SIDS_file.read())
@@ -98,48 +105,32 @@ class Client:
             else:
                 raise AuthError("You haven't authenticated before.")
 
-    def get_raw_contacts(self):
-        """
-        Gets all textnow contacts
-        """
-        params = (
-            ('page_size', '50'),
-        )
-        res = requests.get("https://www.textnow.com/api/v3/contacts", params=params, cookies=self.cookies)
-        contacts = json.loads(res.text)
-        return contacts["result"]
-
-    def get_contacts(self):
-        raw_contacts = self.get_raw_contacts()
-        contact_list = [Contact(contact, self) for contact in raw_contacts]
-        contacts = Container(contact_list, self)
-        return contacts
-
     def get_messages(self):
         """
             This gets most of the messages both sent and received. However It won't get all of them just the past 10-15
         """
-        message_list = [Message(msg, self) if msg["type"] == MESSAGE_TYPE else MultiMediaMessage(msg, self) for msg in self.get_raw_messages()]
-        messages = Container(message_list, self)
-        return messages
+        req = requests.get("https://www.textnow.com/api/users/" + self.username + "/messages", headers=self.headers,
+                           cookies=self.cookies)
+        if str(req.status_code).startswith("2"):
+            messages = json.loads(req.content)
+            messages = [
+                Message(msg, self) if not msg["message"].startswith("http") else MultiMediaMessage(msg, self)
+                for msg in messages["messages"]]
+            return MessageContainer(messages, self)
+        else:
+            self.request_handler(req.status_code)
 
     def get_raw_messages(self):
         """
-            This gets most of the messages both sent and received. It takes about 30 seconds though
+            This gets most of the messages both sent and received. However It won't get all of them just the past 10-15
         """
-        all_messages = []
-        for contact in self.get_contacts():
-            req = self.session.get("https://www.textnow.com/api/users/" + self.username + f"/messages?contact_value={contact.number}&start_message_id=99999999999999&direction=past&page_size=200&get_archived=1", headers=self.headers, cookies=self.cookies)
-            self.session.get(req.url)
-            if req.cookies['connect.sid'] != self._user_sid:
-                print(req.cookies["connect.sid"])
-                self._user_sid = req.cookies["connect.sid"]                   
-            if str(req.status_code).startswith("2"):
-                messages = json.loads(req.content)
-                all_messages.append(messages["messages"])
-            else:
-                self.request_handler(req.status_code)
-        return all_messages
+        req = requests.get("https://www.textnow.com/api/users/" + self.username + "/messages", headers=self.headers,
+                           cookies=self.cookies)
+        if str(req.status_code).startswith("2"):
+            messages = json.loads(req.content)
+            return messages["messages"]
+        else:
+            self.request_handler(req.status_code)
 
     def get_sent_messages(self):
         """
@@ -148,7 +139,7 @@ class Client:
         sent_messages = self.get_messages()
         sent_messages = [msg for msg in sent_messages if msg.direction == SENT_MESSAGE_TYPE]
 
-        return Container(sent_messages, self)
+        return MessageContainer(sent_messages, self)
 
     def get_received_messages(self):
         """
@@ -157,7 +148,7 @@ class Client:
         messages = self.get_messages()
         messages = [msg for msg in messages if msg.direction == RECEIVED_MESSAGE_TYPE]
 
-        return Container(messages, self)
+        return MessageContainer(messages, self)
 
     def get_unread_messages(self):
         """
@@ -166,7 +157,7 @@ class Client:
         new_messages = self.get_received_messages()
         new_messages = [msg for msg in new_messages if not msg.read]
 
-        return Container(new_messages, self)
+        return MessageContainer(new_messages, self)
 
     def get_read_messages(self):
         """
@@ -175,7 +166,7 @@ class Client:
         new_messages = self.get_received_messages()
         new_messages = [msg for msg in new_messages if msg.read]
 
-        return Container(new_messages, self)
+        return MessageContainer(new_messages, self)
 
     def send_mms(self, to, file):
         """
@@ -186,11 +177,8 @@ class Client:
         has_video = True if file_type == "video" else False
         msg_type = 2 if file_type == "image" else 4
 
-        file_url_holder_req = self.session.get("https://www.textnow.com/api/v3/attachment_url?message_type=2",
+        file_url_holder_req = requests.get("https://www.textnow.com/api/v3/attachment_url?message_type=2",
                                            cookies=self.cookies, headers=self.headers)
-        if file_url_holder_req.cookies['connect.sid'] != self._user_sid:
-            print(file_url_holder_req.cookies["connect.sid"])
-            self._user_sid = file_url_holder_req.cookies["connect.sid"]
         if str(file_url_holder_req.status_code).startswith("2"):
             file_url_holder = json.loads(file_url_holder_req.text)["result"]
 
@@ -206,13 +194,8 @@ class Client:
                     "credentials": 'omit'
                 }
 
-                place_file_req = self.session.put(file_url_holder, data=raw, headers=headers_place_file,
+                place_file_req = requests.put(file_url_holder, data=raw, headers=headers_place_file,
                                               cookies=self.cookies)
-
-                self.session.get(place_file_req.url)
-                if place_file_req.cookies['connect.sid'] != self._user_sid:
-                    print(place_file_req.cookies["connect.sid"])
-                    self._user_sid = place_file_req.cookies["connect.sid"]
                 if str(place_file_req.status_code).startswith("2"):
 
                     json_data = {
@@ -227,7 +210,7 @@ class Client:
                         "media_type": file_type
                     }
 
-                    send_file_req = self.session.post("https://www.textnow.com/api/v3/send_attachment", data=json_data,
+                    send_file_req = requests.post("https://www.textnow.com/api/v3/send_attachment", data=json_data,
                                                   headers=self.headers, cookies=self.cookies)
                     return send_file_req
                 else:
@@ -248,12 +231,8 @@ class Client:
                         self.username + '","has_video":false,"new":true,"date":"' + datetime.now().isoformat() + '"} '
             }
 
-        response = self.session.post('https://www.textnow.com/api/users/' + self.username + '/messages',
+        response = requests.post('https://www.textnow.com/api/users/' + self.username + '/messages',
                                  headers=self.headers, cookies=self.cookies, data=data)
-        self.session.get(response.url)
-        if response.cookies['connect.sid'] != self._user_sid:
-            print(response.cookies["connect.sid"])
-            self._user_sid = response.cookies["connect.sid"]
         if not str(response.status_code).startswith("2"):
             self.request_handler(response.status_code)
         return response
@@ -261,7 +240,7 @@ class Client:
     def wait_for_response(self, number, timeout_bool=True):
         for msg in self.get_unread_messages():
             msg.mark_as_read()
-        timeout = datetime.now() + dt.timedelta(minute=10)
+        timeout = datetime.now() + relativedelta(minute=10)
         if not timeout_bool:
             while 1:
                 unread_msgs = self.get_unread_messages()
@@ -291,5 +270,11 @@ class Client:
 
     def request_handler(self, status_code: int):
         status_code = str(status_code)
-        raise FailedRequest(status_code)
+        if status_code == '401':
+            error = FailedRequest(status_code)
+            print(error)
 
+            self.auth_reset()
+            return
+
+        raise FailedRequest(status_code)
