--- pytextnow/TN_objects/user.py
@@ -1,32 +1,0 @@
-from database.db import DatabaseHandler
-
-class User(object):
-
-    def __init__(self, username, sid=None, id=None) -> None:
-        self.id = id
-        self.db_handler = DatabaseHandler()
-        self.sid = sid
-        self.username = username
-        if not self.sid and self.username:
-            self.get_sid(self.username)
-        super().__init__()
-
-    def get_sid(self, username: str):
-        """
-        Get a users SID by username or return None
-        """
-        return self.db_handler._map_to_class(
-                self.db_handler.filter(
-                    "user_sids", {
-                        "username": username
-                    }
-                )[0]
-            ).sid
-    
-    @property
-    def sid(self):
-        """
-        Work around for self.get_sid() throwing an error when
-        we get no sid from the database
-        """
-        return self.sid if len(self.sid) > 0 else None
