--- pytextnow/constants.py
@@ -1,12 +1,0 @@
-MESSAGE_TYPE = 0
-MULTIMEDIA_MESSAGE_TYPE = 1
-
-#new
-USER_TYPE = 3
-CONTACT_TYPE = 4
-
-
-SENT_MESSAGE_TYPE = 2
-RECEIVED_MESSAGE_TYPE = 1
-
-SIP_ENDPOINT = "prod.tncp.textnow.com"
